chore(api): bump dependencies
This includes a cryptography update, which now requires Rust to build,
significantly prolonging build time.

Prebuilt cryptography wheels are not supported on Alpine. It would work
on Ubuntu and Debian, but Ubuntu is not an option because the Python
Docker image has no Ubuntu version, and the Debian version comes with a
Postgres client version that is too old for us. So for now, we're stuck
with Alpine and (now slowly) building cryptography ourselves.